pyproject.toml
[tool.poetry]
name = "naw_tools"
version = "0.1.0"
description = ""
authors = ["Vivien Robert <vivienrobertbarbe@gmail.com>"]
packages = [
    {include = "naw_tools"},
    {include = "naw_utils"}
]

[tool.poetry.dependencies]
python = ">=3.12.0,<3.13"
pandas = "^2.2.2"
scipy = "^1.8.1"
protobuf = "3.20.*"
WTForms = "^3.0.1"
loguru = "^0.6.0"
click = "^8.1.7"
regex = "^2024.7.24"

[tool.poetry.group.flask_app]
optional = true

[tool.poetry.group.flask_app.dependencies]
Flask = "^2.0.1"
Flask-WTF = "^0.15.1"
Flask-Bootstrap4 = "^4.0.2"

[tool.poetry.group.gradio_app.dependencies]
gradio = "^4.38.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.1"
hypothesis = "^6.108.4"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
